Pulled dtypes out into separate module
